+++ src/pages/registro/RegistroPage.jsx
@@ -1,4 +1,4 @@
-import { registrar } from "../../api/api";
+import { registrar } from "../../services/UsuarioLinkedServices";
 import { useState } from "react";
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
