+++ src/services/StatPageServices.jsx
@@ -1,0 +1,11 @@
+import axios from "axios";
+import {api} from "../api/api";
+
+export const getEstadisticas = async () => {
+  try {
+    const response = await axios.get(`${api.defaults.baseURL}/estadisticas`);
+    return response.data;
+  } catch (error) {    
+    throw error;
+  }
+};
